pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-test"
version = "0.1.0"
description = "MCP Tic-Tac-Toe with Ollama"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "httpx>=0.25.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
]

[project.scripts]
mcp-test = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["mcp_server", "mcp_client", "web_ui"]